Fazendo alteração dos ids de Integer para Long e concluindo criação do service e resource de pedido. Aplicação funcionando por requisição rest

+++ curso-spring-restapi/src/main/java/com/jpbandeira/springrestapi/repositories/ItemPedidoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ItemPedidoRepository extends JpaRepository<ItemPedido, Integer>{
+public interface ItemPedidoRepository extends JpaRepository<ItemPedido, Long>{
 
 }
